Match-id-9a4a960956a66f00490739f3b205fcd10f699c6e

+++ csrc/framework/client_parser.cpp
@@ -154,7 +154,7 @@ static void ParseSelectSteps(const std::string &param, UserCommand &userCommand)
             if (!std::regex_match(step, numberPattern)) {
                 return parseFailed();
             }
-            stepListInfo.stepIdList[stepListInfo.stepCount] = stoi(it->str());
+            stepListInfo.stepIdList[stepListInfo.stepCount] = static_cast<uint32_t>(stoi(it->str()));
             stepListInfo.stepCount++;
         }
